refactor(scout-for-lol): move explore-mention-picker into the explore sub-domain
Why:
main's directory-split refactor (22ec6d32c7) moved every explore-prefixed
component under components/explore/. This branch's rebase re-added
explore-mention-picker.tsx at the old top-level path, which both breaks the
convention and adds back to the directory the split was meant to shrink.

What:
git mv into components/explore/ and fix the one importer.

Verification:
bunx turbo run lint typecheck --filter=@scout-for-lol/app (run after full stack rebase).
