+++ packages/scout-for-lol/packages/app/src/components/explore/explore-mention-picker.tsx
